# perf(modeling): convert {} to Map in insertTjunctions