Followup from #1051

This PR replaces js objects `{}` with faster `Map` in insertTjunctions. There is a measurable speed improvement depending on the model. For `generalize` applied to two unioned spheres, this results in improvements:

 - generalize: 188ms -> 96ms
 - insertTjunctions (step1): 32ms -> 32ms
 - insertTjunctions (step2): 3ms -> 2ms
 - insertTjunctions (step3): 150ms -> 57ms

```js
const jscad = require('@jscad/modeling')
const { union } = jscad.booleans
const { sphere } = jscad.primitives
const { generalize } = jscad.modifiers

const segments = 100

const main = () => {
  let out = union(sphere({ segments }), sphere({ segments, center: [1, 0, 0] }))
  const startTime = performance.now()
  out = generalize({ triangulate: true }, out)
  console.log(`Generalize time ${performance.now() - startTime} ms`)
  return out
}

module.exports = { main }
```


### All Submissions:

* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../pulls) for the same update/change?
* [x] Does your submission pass tests?